(#237) feat: index 파일 통합
